Break out new ownership panel into its own component
